Cleanup snapshot-name passing and handling
- add snapshotname arg that gets passed by coordinator to worker
- remove extraneous mapping APIs for getting names from IDs
- clean up snapshot deletion test that broke as result
  - implement v2 call for maximum cleanliness

+++ projects/coordinator/src/main/java/org/batfish/coordinator/WorkMgr.java
@@ -671,9 +671,16 @@ public class WorkMgr extends AbstractCoordinator {
     return true;
   }
 
-  public void delSnapshot(String network, String snapshot) {
+  public boolean delSnapshot(String network, String snapshot) {
+    if (!_idManager.hasNetworkId(network)) {
+      return false;
+    }
     NetworkId networkId = _idManager.getNetworkId(network);
+    if (!_idManager.hasSnapshotId(snapshot, networkId)) {
+      return false;
+    }
     _idManager.deleteSnapshot(snapshot, networkId);
+    return true;
   }
 
   public void delQuestion(String network, String qName) {
@@ -1847,6 +1854,7 @@ public class WorkMgr extends AbstractCoordinator {
     }
     SnapshotId snapshotId = idManager.getSnapshotId(snapshot, networkId);
     params.put(BfConsts.ARG_TESTRIG, snapshotId.getId());
+    params.put(BfConsts.ARG_SNAPSHOT_NAME, snapshot);
 
     // referenceSnapshot
     String referenceSnapshot = params.get(BfConsts.ARG_DELTA_TESTRIG);
